Fixed thrown exception when calling dataRetriever after receiving a CacheUpdated message
This bug was found when our IIS instances were continually restarting. For
some reason CacheUpdated messages were being sent out but the data was not
in Redis. In that case, the null dataRetriever was being called which would
throw a NullReferenceException. Since it's running in an EventHandler in
the background it wasn't being caught and the process would crash. Changed
the null Task to a Task that returns null so we don't throw an exception.